Cygwin: add ability to pass cwd to child process
This will be used by posix_spawn_fileaction_add_(f)chdir.

The int cwdfd is placed such that it fits into space previously unused
due to alignment in the cygheap_exec_info class.

This uses a file descriptor rather than a path both because it is easier
to marshal to the child and because this should protect against races
where the directory might be renamed or removed between addfchdir and
the actual setting of the cwd in the child.

Signed-off-by: Jeremy Drake <cygwin@jdrake.com>