Cygwin: fix fcntl F_GETLK
The commit implementing OFD locks dropped the F_GETLK case from the
switch in fhandler_base::lock, replacing it with F_OFD_GETLK.  This
appears to have been an oversight, as F_OFD_SETLK was added as an
additional case above.

This resulted in the winsup.api/ltp/fcntl05 test failing.

Fixes: a66ed519884d ("Cygwin: fcntl: implement Open File Description (OFD) locks")
Signed-off-by: Jeremy Drake <cygwin@jdrake.com>